totes solved the foreign key issue. still errs tho
I didn't even read anything, I just solved it. For some reason it was coming in as a weird object that was like a single-valued object in an array of length 1.

My regex is having issues parsing some dates though, so I need to see what is causing that issue.