Works like a charm but for one unacceptable thing.
I thought every class had dates, but it turns out that some of them are TBA. I suppose this is a lesson. I was trying to make a very rigid structure in which to hold data that could change. Perhaps I should 'loosen up' my database restrictions and node code a little bit.