Did some stuff for MVP that I will surely trash.
I decided to not use Jade and to switch to an API-first client-server model for both mobile and web. Of course it is already this way for mobile, but now I will use AngularJS as the client for web, which will call my API to load page content.

My Express server will no longer have to think about the webpage, and will just do what is asked.